feat: refactor doctor dashboard and related components for improved state management and type definitions
