+++ services/doctor/types.ts
@@ -18,64 +18,6 @@ export interface DoctorUpdateRequest {
   // Thêm các trường khác nếu cần
 }
 
-export interface DoctorAppointment {
-  appointmentID: string;
-  patient: {
-    patientID: string;
-    account: {
-      id: string;
-      username: string;
-      email: string;
-      phone: string;
-      avatar: string;
-      role: string;
-      isDeleted: boolean;
-    };
-    name: string;
-    dob: string;
-    gender: string;
-    medNo: string;
-    medDate: string;
-    medFac: string;
-  };
-  doctor: {
-    doctorID: string;
-    account: {
-      id: string;
-      username: string;
-      email: string;
-      phone: string;
-      avatar: string;
-      role: string;
-      isDeleted: boolean;
-    };
-    name: string;
-    gender: string;
-    specialty: string;
-    degrees: string;
-  };
-  medicalService: {
-    serviceID: number;
-    name: string;
-    description: string;
-    price: number;
-    img: string;
-    type: string;
-    specialty: string;
-    isActive: boolean;
-    isRequireDoctor: boolean;
-    isOnline: boolean;
-    createAt: string;
-  };
-  appointmentDate: string; // ISO date string
-  status: string;
-  isAnonymous: boolean;
-  urlLink: string;
-  note: string;
-  paymentStatus: string | null;
-  createAt: string; // ISO date string
-}
-
 export interface Appointment {
   appointmentID: string;
   patientName: string;
@@ -114,12 +56,4 @@ export interface LabResult {
   referenceRange: string;
   testDate: string;
   performedBy: string;
-}
-
-export interface MedicalRecord {
-  recordId: string;
-  appointmentId: string;
-  diagnosis: string;
-  createDate: string;
-  note: string;
 }
