+++ services/transaction/api.ts
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
 import axiosInstance from "@/config/axios";
 import { Transaction } from "./types";
